// src/main/resources/angular/src/app/api/model/availablePermission.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: kfiertek@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AvailablePermission { 
    permissionId: number;
    name: string;
    vehicleId: number;
    totalQuantity?: number;
    availableQuantity: number;
    availableForUse: boolean;
    /**
     * Description of schedule
     */
    scheduleDescription?: string;
}